🐛 fix(command-palette): refactor getJapaneseReadings function into a separate utility file and update imports
