+++ browser-features/chrome/common/command-palette/multiInputCommand/switcher/closed-window-switcher.ts
@@ -5,25 +5,7 @@ import type {
   PaletteCommand,
   CommandStepChoice,
 } from "../../types.ts";
-
-/**
- * Get hiragana reading keywords for a given action/command ID from i18n.
- * Returns an empty array for non-Japanese locales or if no readings are defined.
- */
-function getJapaneseReadings(id: string): string[] {
-  try {
-    const readings: unknown = i18next.t(`commandPaletteReadings.${id}`, {
-      defaultValue: [] as string[],
-      returnObjects: true,
-    });
-    if (Array.isArray(readings)) {
-      return readings.filter((r): r is string => typeof r === "string");
-    }
-    return [];
-  } catch {
-    return [];
-  }
-}
+import { getJapaneseReadings } from "../../utils/getJapaneseReadings.ts";
 
 export function loadClosedWindows(): Promise<CommandStepChoice[]> {
   try {
