+++ browser-features/chrome/common/command-palette/multiInputCommand/switcher/history-switcher.ts
@@ -13,26 +13,7 @@ import type {
   SqliteModule,
   HistoryPlacesUtilsModule,
 } from "./types.ts";
-
-/**
- * Get hiragana reading keywords for a given action/command ID from i18n.
- * Returns an empty array for non-Japanese locales or if no readings are defined.
- */
-function getJapaneseReadings(id: string): string[] {
-  try {
-    const readings: unknown = i18next.t(`commandPaletteReadings.${id}`, {
-      defaultValue: [] as string[],
-      returnObjects: true,
-    });
-    if (Array.isArray(readings)) {
-      return readings.filter((r): r is string => typeof r === "string");
-    }
-    return [];
-  } catch (e) {
-    console.error("[HistorySwitcher]", "getJapaneseReadings i18n error:", e);
-    return [];
-  }
-}
+import { getJapaneseReadings } from "../../utils/getJapaneseReadings.ts";
 
 const PAGE_SIZE = 20;
