+++ browser-features/chrome/common/command-palette/multiInputCommand/switcher/tab-switcher.ts
@@ -6,25 +6,7 @@ import type {
   CommandStepChoice,
   StepChoicesResult,
 } from "../../types.ts";
-
-/**
- * Get hiragana reading keywords for a given action/command ID from i18n.
- * Returns an empty array for non-Japanese locales or if no readings are defined.
- */
-function getJapaneseReadings(id: string): string[] {
-  try {
-    const readings: unknown = i18next.t(`commandPaletteReadings.${id}`, {
-      defaultValue: [] as string[],
-      returnObjects: true,
-    });
-    if (Array.isArray(readings)) {
-      return readings.filter((r): r is string => typeof r === "string");
-    }
-    return [];
-  } catch {
-    return [];
-  }
-}
+import { getJapaneseReadings } from "../../utils/getJapaneseReadings.ts";
 
 interface ChromeWindow extends Window {
   gBrowser?: GBrowser;
